+++ src/util/scanAdapters/EtherscanV2ScanAdapter.ts
@@ -51,9 +51,34 @@ export function makeEtherscanV2ScanAdapter(
     }
     const data = await response.json()
 
-    // console.log(data)
-
     if (data.status === '1' && data.result.length > 0) {
+      return true
+    }
+
+    // If no normal transactions, check for token transactions:
+    const tokenParams = new URLSearchParams({
+      chainId: chainId.toString(),
+      module: 'account',
+      action: 'tokentx',
+      address: normalizedAddress,
+      startblock: checkpoint,
+      endblock: '999999999',
+      sort: 'asc'
+    })
+    if (apiKey != null) {
+      tokenParams.set('apikey', apiKey)
+    }
+    const tokenResponse = await fetch(`${url}/v2/api?${tokenParams.toString()}`)
+    if (tokenResponse.status !== 200) {
+      logger.error(
+        'scanAddress tokenTx error',
+        tokenResponse.status,
+        tokenResponse.statusText
+      )
+      return false
+    }
+    const tokenData = await tokenResponse.json()
+    if (tokenData.status === '1' && tokenData.result.length > 0) {
       return true
     }