+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: Added support for `scanAddress` API for EvmRpc plugin by using etherscan services.
+
 ## 0.1.1 (2025-04-29)
 
 - fixed: Fix subscribe method spec for change protocol.